requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
Pillow>=10.0          # For ImageField (company logo)
